Split middleware into separate files
