+++ src/middleware/with-saleor-domain-present.ts
@@ -1,0 +1,17 @@
+import { Middleware } from "retes";
+import { Response } from "retes/response";
+
+import { getSaleorHeaders } from "../headers";
+
+export const withSaleorDomainPresent: Middleware = (handler) => async (request) => {
+  const { domain } = getSaleorHeaders(request.headers);
+
+  if (!domain) {
+    return Response.BadRequest({
+      success: false,
+      message: "Missing Saleor domain header.",
+    });
+  }
+
+  return handler(request);
+};
